Began working on Cut the Wheat Microgame
Created new scene folder in Assets folder for Cut the Wheat. Created new scene for microgame, set up the scene with basic geometry, added script file for CursorFollower. Commiting here as I need to go to main to create a new Input Action Asset map for the input of this game to hopefully avoid conflicts down the line