Completed (sort of) SheepController + AnimalController functionality
The block representing the sheep in the scene now moves left and right. The script is still semi unfinishd as it depends on functions in scripts that have not been written as of now. For now, comments have been left in theit place, as well as a temporary bool used in place for the main functionality to work