Imported sheep model + apple icons
Animation file for testing how to add animations to the sheep is in project but have not gotten to work

Undo of previous commit

Hoping this works!

Revert "Undo of previous commit"

This reverts commit 61d8b07841ea9fc6ceb5590ca8d0025af809d45f.

Removed sitting animation

Done as github cant hold file size of that. Hoping to squash this with the original imported commit so I can push this to the online repo