Attempted to resolve bug with Gronknolius Not Moving Correctly
Gronknolius will jitter in the cnetre while ReadjustPosition and Rotate() are spammed. This is due to something with Collider.bounds.Contains(), which appears to be giving the wrong information. WHen the collider we call this from is moved above Gronk, he will begin moving as wanted. When the cllider is brought back down, the same thing repeats