Added Animal Icons
Not actually added to the gameobject but imported in the asset folder and made into a sprite so I can export it as a unity pacakge into develop. Im doing this primarily so as not to cause a merge conflict for Elias who'll later merge in his stuff. Creating a sprite in develop would install the 2D packages which Elias doesn't have on his end. As today's the deadline, we we want to prevent any conflicts as much as possible